move cqlsh copy tests to their own file
